Split Replacement into own package
Also requires moving ParseVersion to util
