+++ src/bci_build/package/apache_tomcat.py
@@ -9,8 +9,8 @@ from bci_build.package import DOCKERFILE_RUN
 from bci_build.package import ApplicationStackContainer
 from bci_build.package import OsContainer
 from bci_build.package import Package
-from bci_build.package import Replacement
 from bci_build.package import _build_tag_prefix
+from bci_build.replacement import Replacement
 
 # last version needs to be the newest
 _TOMCAT_VERSIONS: list[str] = ["9", "10.1"]
