+++ src/bci_build/package/appcontainers.py
@@ -14,13 +14,13 @@ from bci_build.package import DOCKERFILE_RUN
 from bci_build.package import ApplicationStackContainer
 from bci_build.package import OsContainer
 from bci_build.package import Package
-from bci_build.package import ParseVersion
-from bci_build.package import Replacement
 from bci_build.package import _build_tag_prefix
 from bci_build.package.helpers import generate_from_image_tag
 from bci_build.package.helpers import generate_package_version_check
 from bci_build.package.versions import format_version
 from bci_build.package.versions import get_pkg_version
+from bci_build.replacement import Replacement
+from bci_build.util import ParseVersion
 
 
 def _envsubst_pkg_name(os_version: OsVersion) -> str:
