+++ src/bci_build/package/kubevirt_cdi.py
@@ -8,13 +8,13 @@ from bci_build.containercrate import ContainerCrate
 from bci_build.os_version import OsVersion
 from bci_build.package import DOCKERFILE_RUN
 from bci_build.package import ApplicationStackContainer
-from bci_build.package import ParseVersion
-from bci_build.package import Replacement
 from bci_build.package.helpers import generate_from_image_tag
 from bci_build.package.helpers import generate_package_version_check
 from bci_build.package.kubevirt import KubeVirtRegistry
 from bci_build.package.versions import format_version
 from bci_build.package.versions import get_pkg_version
+from bci_build.replacement import Replacement
+from bci_build.util import ParseVersion
 
 CDI_EXCLUSIVE_ARCH = [Arch.X86_64]
