+++ src/bci_build/package/openjdk.py
@@ -11,9 +11,9 @@ from bci_build.os_version import _SUPPORTED_UNTIL_SLE
 from bci_build.os_version import OsVersion
 from bci_build.package import DOCKERFILE_RUN
 from bci_build.package import DevelopmentContainer
-from bci_build.package import Replacement
 from bci_build.package import _build_tag_prefix
 from bci_build.package import generate_disk_size_constraints
+from bci_build.replacement import Replacement
 
 
 def supported_until(os_version: OsVersion, jre_major: int) -> datetime.date | None:
