+++ src/bci_build/package/pulseaudio.py
@@ -11,13 +11,13 @@ from bci_build.os_version import OsVersion
 from bci_build.package import DOCKERFILE_RUN
 from bci_build.package import ApplicationStackContainer
 from bci_build.package import OsContainer
-from bci_build.package import ParseVersion
-from bci_build.package import Replacement
 from bci_build.package import _build_tag_prefix
 from bci_build.package.helpers import generate_package_version_check
 from bci_build.package.kiosk import KIOSK_EXCLUSIVE_ARCH
 from bci_build.package.kiosk import KIOSK_SUPPORT_ENDS
 from bci_build.package.kiosk import KioskRegistry
+from bci_build.replacement import Replacement
+from bci_build.util import ParseVersion
 
 _PULSE_FILES = {
     "client.conf": (
