+++ src/bci_build/package/trivy.py
@@ -6,9 +6,9 @@ from bci_build.os_version import CAN_BE_LATEST_OS_VERSION
 from bci_build.os_version import OsVersion
 from bci_build.package import ApplicationStackContainer
 from bci_build.package import Package
-from bci_build.package import ParseVersion
-from bci_build.package import Replacement
 from bci_build.package.helpers import generate_from_image_tag
+from bci_build.replacement import Replacement
+from bci_build.util import ParseVersion
 
 _TRIVY_VER_RE = "%%trivy_version%%"
