+++ src/bci_build/package/valkey.py
@@ -9,12 +9,12 @@ from bci_build.os_version import CAN_BE_LATEST_OS_VERSION
 from bci_build.package import DOCKERFILE_RUN
 from bci_build.package import ApplicationStackContainer
 from bci_build.package import OsContainer
-from bci_build.package import ParseVersion
-from bci_build.package import Replacement
 from bci_build.package import _build_tag_prefix
 from bci_build.package.helpers import generate_package_version_check
 from bci_build.package.versions import format_version
 from bci_build.package.versions import get_pkg_version
+from bci_build.replacement import Replacement
+from bci_build.util import ParseVersion
 
 VALKEY_CONTAINERS = [
     ApplicationStackContainer(
